requirements.txt
# Core dependencies
Flask==2.3.3
python-dotenv==1.0.0

# Data processing
pandas>=2.0.0
sqlalchemy>=2.0.0

# Database drivers
psycopg2-binary>=2.9.0
mysql-connector-python>=8.0.0
presto-python-client>=0.8.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.0.0

# Template and reporting
jinja2>=3.1.0

# AI/ML integrations (optional)
openai>=1.0.0
anthropic>=0.64.0

# Progress and CLI
tqdm>=4.65.0
click>=8.1.0